# contracts/openapi/app-api.yaml
openapi: 3.1.0
info:
  title: Goat OS App API
  version: 0.1.0
  summary: Goat OS app and operating API for field tasks, goat lookup, protocol config, vaccination, and process integrity.
  description: REST/JSON API for operator/mobile and current admin-web operating surfaces. Reads and writes are RBAC
    scope-filtered before response DTOs are built.
servers:
  - url: https://api.goatos.example.invalid
    description: Synthetic contract server placeholder.
security:
  - bearerAuth: []
...
paths:
...
  /protocols/vaccination/impact-preview:
    post:
      tags:
        - Vaccination
      operationId: vaccinationImpactPreview
      summary: Aggregate config impact preview (eligible animals, vaccination cells, affected sheds, estimated days) from the eligibility rollup read model.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImpactPreviewInput"
      responses:
        "200":
          description: Vaccination impact preview.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpactPreviewResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /vaccination/verification-queue:
    get:
      tags:
        - Vaccination
      operationId: vaccinationVerificationQueue
      summary: List recorded vaccination completions awaiting proof verification.
      parameters:
        - name: park_id
          in: query
          required: false
          description: Optional park scope (top-bar park); filters the queue to that park's completions.
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: cursor
          in: query
          required: false
          description: Opaque keyset cursor returned by the previous verification-queue page.
          schema:
            type: string
      responses:
        "200":
          description: Vaccination verification queue.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaccinationQueueResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Permission denied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - field_errors
        - trace_id
        - retryable
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 120
        message:
          type: string
          minLength: 1
          maxLength: 500
        field_errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        trace_id:
          type: string
          minLength: 1
          maxLength: 200
        retryable:
          type: boolean
    FieldError:
      type: object
      additionalProperties: false
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 200
        code:
          type: string
          minLength: 1
          maxLength: 120
        message:
          type: string
          minLength: 1
          maxLength: 500
...
    WorkState:
      type: string
      enum:
        - scheduled
        - due
        - overdue
        - in_progress
        - proof_pending
        - verification_pending
        - rejected
        - deferred
        - missed
        - blocked
        - completed
...
    VaccinationQueueItem:
      type: object
      additionalProperties: false
      required:
        - completion_id
        - obligation_id
        - goat_id
        - administered_at
        - doses
        - work_state
      properties:
        completion_id:
          type: string
        obligation_id:
          type: string
        goat_id:
          type: string
        batch_id:
          type: string
        sop_task_id:
          type: string
        sop_task_row_version:
          type: integer
          minimum: 1
        administered_at:
          type: string
          format: date-time
        doses:
          type: integer
        route_site:
          type: string
        work_state:
          $ref: "#/components/schemas/WorkState"
    VaccinationQueueResponse:
      type: object
      additionalProperties: false
      required:
        - items
        - total_count
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VaccinationQueueItem"
        total_count:
          type: integer
          format: int64
        next_cursor:
          type: string
    ImpactPreviewInput:
      type: object
      additionalProperties: false
      description: >-
        Config impact-preview filter. Empty filter dims mean "any". Preview is aggregate-only over the
        eligibility rollup, so there is no per-goat dose count and no time-relative warmup on this path.
      properties:
        species:
          type: string
        stage:
          type: string
        sex:
          type: string
        breed:
          type: string
        health:
          type: string
        park_id:
          type: string
        vaccine_item_id:
          type: string
        location_id:
          type: string
        dose_rows:
          type: integer
          minimum: 0
          description: Number of selected dose/schedule rows (vaccination cells per eligible animal).
        daily_cap:
          type: integer
          minimum: 1
          description: >-
            Draft daily vaccination cap (vaccinations/day) authored in the rule editor, used to compute
            estimated_days BEFORE publish. Omit to fall back to the published/operational cap.
        max_buffer_days:
          type: integer
          minimum: 0
          description: >-
            Draft safe-window buffer (days) authored in the rule editor. Omit to fall back to the
            business default. Drives the within_cap / over_cap (split) / capacity_breach classification.
        horizon_days:
          type: integer
          minimum: 0
    ImpactPreviewResult:
      type: object
      additionalProperties: false
      description: >-
        Aggregate-only config impact preview. Every number is read from the precomputed vaccination
        eligibility rollup (read model); the request path never scans goats. Use the business names
        only (eligible_animals / vaccination_cells / affected_sheds / estimated_days / daily_cap).
      required:
        - eligible_animals
        - vaccination_cells
        - affected_sheds
        - estimated_days
        - daily_cap
        - planned_sessions
        - source_revision
        - warnings
      properties:
        eligible_animals:
          type: integer
          description: SUM(animal_count) over usable animals matching the filter.
        vaccination_cells:
          type: integer
          description: eligible_animals × selected dose rows.
        affected_sheds:
          type: integer
          description: Distinct sheds holding usable animals in scope.
        estimated_days:
          type: integer
          description: ceil(vaccination_cells / daily_cap).
        daily_cap:
          type: integer
          description: Configured vaccinations/day used for estimated_days.
        capacity_status:
          type: string
          enum: [within_cap, over_cap, capacity_breach]
          description: >-
            Planner classification of the draft under daily_cap + the buffer window: within_cap (fits one
            day), over_cap (Split — fits the safe window), capacity_breach (Needs review — beyond it).
            Omitted when there are no cells to plan.
        planned_sessions:
          type: array
          description: >-
            Pre-publish session split under the draft daily cap. Bounded for very large herds; estimated_days
            remains the full duration when the preview is truncated.
          items:
            $ref: "#/components/schemas/VaccinationPlannedSession"
        doses_available:
          type: string
          description: Optional cheap stock check; present only when a vaccine item is set.
        earliest_expiry:
          type: string
          format: date-time
        source_revision:
          type: integer
          format: int64
          description: Read-model recompute stamp behind these numbers (0 when the rollup has no data for the scope yet).
        recomputed_at:
          type: string
          format: date-time
          description: When the read-model grain behind these numbers was last rebuilt.
        warnings:
          type: array
          items:
            type: string
...
    VaccinationCapacityStatus:
      type: string
      enum:
        - within_cap
        - over_cap
        - capacity_breach
      description: |
        Machine vocabulary for capacity/session-splitting state:
        - within_cap: fits in a single day (sessions <= 1)
        - over_cap: safely split across multiple days within the safe window
        - capacity_breach: cannot fit within the safe window; needs manager review

        NEVER show raw tokens in CEO UI. Frontend renders provided labels:
        within_cap -> "Within cap", over_cap -> "Split", capacity_breach -> "Needs review"
    VaccinationPlannedSession:
      type: object
      additionalProperties: false
      required:
        - date
        - vaccinations
        - dailyLimit
        - capacity
      properties:
        date:
          type: string
          description: Asia/Kolkata business date (YYYY-MM-DD)
        vaccinations:
          type: integer
          description: Vaccination cells (obligations) planned for this day
        dailyLimit:
          type: integer
          description: The configured daily vaccination capacity cap
        capacity:
          $ref: "#/components/schemas/VaccinationCapacityStatus"
...